Move events into subject folder (#35262)
