+++ middleware/req-utils.js
@@ -1,4 +1,4 @@
-import Hydro from '../lib/hydro.js'
+import Hydro from '../src/events/hydro.js'
 
 export default function reqUtils(req, res, next) {
   req.hydro = new Hydro()
