+++ nodemon.json
@@ -6,6 +6,7 @@
     "stylesheets",
     "tests",
     "content",
+    "translations",
     "data/reusables",
     "data/variables",
     "data/glossaries",
