+++ script/create-subject.js
@@ -7,14 +7,15 @@ Output looks like:
 src/
   xsubject/
     README.md
-    components/
+    docs/
       gitkeep
     lib/
-    tests/
-    scripts/
-    stylesheets/
+    middleware/
     pages/
-    docs/
+    components/
+    stylesheets/
+    scripts/
+    tests/
 */
 
 import fs from 'fs/promises'
@@ -33,7 +34,16 @@ if (!name) {
 
 const src = 'src/'
 
-const subfolders = ['components', 'lib', 'tests', 'scripts', 'stylesheets', 'pages', 'docs']
+const subfolders = [
+  'docs',
+  'lib',
+  'middleware',
+  'pages',
+  'components',
+  'stylesheets',
+  'scripts',
+  'tests',
+]
 
 const files = [
   [
