+++ src/events/tests/events.js
@@ -1,5 +1,5 @@
 import { expect, jest } from '@jest/globals'
-import { post } from '../helpers/e2etest.js'
+import { post } from '../../../tests/helpers/e2etest.js'
 
 describe('POST /events', () => {
   jest.setTimeout(60 * 1000)
